+++ packages/hono/src/response/error/http.ts
@@ -1,0 +1,5 @@
+import { HTTPException } from "hono/http-exception";
+
+class JderHttpException extends HTTPException {}
+
+export { JderHttpException };