+++ packages/hono/tsdown.config.ts
@@ -1,11 +1,12 @@
-import type { Options } from "tsdown";
+import type { UserConfig } from "tsdown";
 
 import { defineConfig } from "tsdown";
 
-const options: Options = {
+const options: UserConfig = {
     entry: {
         response: "./src/response/index.ts",
-        "response/error": "./src/response/error/index.ts",
+        "response/error/index": "./src/response/error/index.ts",
+        "response/error/http": "./src/response/error/http.ts",
         "body-limit": "./src/middlewares/body-limit.ts",
         "ip-limit": "./src/middlewares/ip-limit.ts",
         "time-limit": "./src/middlewares/time-limit.ts",