+++ packages/hono-openapi/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1 (2025-11-25)
+
+### What's Changed
+
+- Update dependencies
+
 ## 0.1.0 (2025-10-29)
 
 First release