+++ packages/hono-standard-validator/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0 (2025-11-25)
+
+### What's Changed
+
+- Use `JderHttpException` which extends `HttpException` from Hono
+
 ## 0.1.0 (2025-10-29)
 
 First release