+++ packages/hono-zod-validator/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.7.0 (2025-11-25)
+
+### What's Changed
+
+- Use `JderHttpException` which extends `HttpException` from Hono
+
 ## 0.6.0 (2025-09-20)
 
 ### Breaking Changes